backend/requirements.txt
# Core dependencies - install first
numpy==1.24.3
pillow==10.0.1

# Flask and web dependencies
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7

# Face recognition dependencies - install with longer timeout
dlib==19.24.2
face-recognition==1.3.0
opencv-python==4.8.1.78

# Additional utilities
python-multipart==0.0.6
